Change factories to methods instead of classes
There is no need for the factories to be classes.
