+++ fee_calculator/apps/viewer/tests/presenters/test_offence_class_presenters.py
@@ -2,7 +2,35 @@ from django.test import TestCase
 
 from calculator.models import OffenceClass
 from viewer.presenters.offence_class_presenters import (
-    AlphaOffenceClassPresenter, NumericOffenceClassPresenter, NoneOffenceClassPresenter, NullOffenceClassPresenter)
+    offence_class_presenter_factory_from_pk,
+    AlphaOffenceClassPresenter,
+    NumericOffenceClassPresenter,
+    NoneOffenceClassPresenter,
+    NullOffenceClassPresenter
+)
+
+
+class OffenceClassPresenterFactoryTestCase(TestCase):
+    def test_creates_none_presenter_for_none_offence_class(self):
+        self.assertIsInstance(offence_class_presenter_factory_from_pk('None'), NoneOffenceClassPresenter)
+
+    def test_creates_none_presenter_for_id_none(self):
+        self.assertIsInstance(offence_class_presenter_factory_from_pk(None), NoneOffenceClassPresenter)
+
+    def test_creates_null_presenter_for_blank_id(self):
+        self.assertIsInstance(offence_class_presenter_factory_from_pk(''), NullOffenceClassPresenter)
+
+    def test_creates_alpha_presenter_for_known_offence_class(self):
+        OffenceClass.objects.create(id='M')
+        presenter = offence_class_presenter_factory_from_pk('M')
+        self.assertIsInstance(presenter, AlphaOffenceClassPresenter)
+        self.assertEqual(presenter.label, 'M')
+
+    def test_creates_numeric_presenter_for_known_offence_class(self):
+        OffenceClass.objects.create(id='50.5')
+        presenter = offence_class_presenter_factory_from_pk('50.5')
+        self.assertIsInstance(presenter, NumericOffenceClassPresenter)
+        self.assertEqual(presenter.label, '50.5')
 
 
 class AlphaOffenceClassPresenterTestCase(TestCase):
